% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPOPT.R
\name{IPOPT}
\alias{IPOPT}
\title{Use Ipopt solver to solve optimization problems.}
\usage{
IPOPT(x, x_L, x_U, g_L, g_U, eval_f, eval_g, eval_grad_f, jac_g1, jac_g2, h1,
  h2)
}
\arguments{
\item{x}{Starting point}

\item{x_L}{Variable lower bounds}

\item{x_U}{Variable upper bounds}

\item{g_L}{Constraint lower bounds}

\item{g_U}{Constraint upper bounds}

\item{eval_f}{Callback: objective function}

\item{eval_g}{Callback: constraint evaluation}

\item{eval_grad_f}{Callback: objective function gradient}

\item{jac_g1, jac_g2}{Callback: Jacobian evaluation}

\item{h1, h2}{Callback: Hessian evaluation}
}
\value{
The return value will be a list which contains three components:
  status: the status of the ipopt solver,
  value: the optimized objective value,
  x: the best set of parameters found.
}
\description{
\code{IPOPT} returns the solution to the optimization problem found by Ipopt solver.
}
