% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPOPT.R
\name{ipopt_setup}
\alias{ipopt_setup}
\title{Do initial setup for ipoptjlr package.}
\usage{
ipopt_setup(...)
}
\arguments{
\item{...}{arguments passed to \code{JuliaCall::julia_setup}.}
}
\description{
\code{ipopt_setup} does the initial setup for ipoptjlr package.
}
\examples{
\dontrun{
ipopt_setup()
}

}
