% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPOPT.R
\name{setup}
\alias{setup}
\title{Do initial setup for the ipoptjlr package.}
\usage{
setup(startup_safe = FALSE)
}
\arguments{
\item{startup_safe}{if you would like to use setup in your .Rprofile, you'd better set startup_safe to TRUE.}
}
\description{
\code{setup} does the initial setup for the ipoptjlr package.
}
\examples{
setup()

}
