% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPOPT.R
\name{ipopt_setup}
\alias{ipopt_setup}
\title{Do initial setup for the ipoptjlr package.}
\usage{
ipopt_setup()
}
\description{
\code{setup} does the initial setup for the ipoptjlr package.
}
\examples{
\dontrun{
ipopt_setup()
}

}
